gnus-summary-article-header for data; org-gnus-follow-link to open
* use gnus-summary-article-header for getting article data

use the gnus-summary-article-header macro to get the article data.

* open gnus-recent articles with org-gnus-follow-link

gnus-recent--article-open now uses the function org-gnus-follow-link, which
simplifies gnus-recent--open-article and is more robust.
Adds an org-gnus dependency.

* ISO date string

the date string in the article title now uses a ISO format similar to the org-mode
date format. It is significantly more compact than the previous US/English style
date format and also can be used for sorting.

* new functions gnus-recent-date-format, gnus-recent-forget-all

  - function gnus-recent-date-format converts the article date to the ISO/org date format use in
    the article title.
  - function gnus-recent-forget-all removes all entries from the
    gnus-recent--article-list. The user can apply it to start over. Added as an
    option to the ivy, helm interfaces.